requirements.txt
flask
scikit-learn
gunicorn
groq
nltk
python-dotenv